+++ src/Repositories/SqlStatementRepository.php
@@ -43,103 +43,103 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
         SqlStatementKeys::CATEGORIES =>
             'SELECT t0.*,
                     (SELECT `value`
-                       FROM eav_attribute t1, catalog_category_entity_varchar t2
+                       FROM ${table:eav_attribute} t1, ${table:catalog_category_entity_varchar} t2
                       WHERE t1.attribute_code = \'name\'
                         AND t1.entity_type_id = 3
                         AND t2.attribute_id = t1.attribute_id
                         AND t2.store_id = 0
                         AND t2.row_id = t0.row_id) AS name,
                     (SELECT `value`
-                       FROM eav_attribute t1, catalog_category_entity_varchar t2
+                       FROM ${table:eav_attribute} t1, ${table:catalog_category_entity_varchar} t2
                       WHERE t1.attribute_code = \'url_key\'
                         AND t1.entity_type_id = 3
                         AND t2.attribute_id = t1.attribute_id
                         AND t2.store_id = 0
                         AND t2.row_id = t0.row_id) AS url_key,
                     (SELECT `value`
-                       FROM eav_attribute t1, catalog_category_entity_varchar t2
+                       FROM ${table:eav_attribute t1}, ${table:catalog_category_entity_varchar} t2
                       WHERE t1.attribute_code = \'url_path\'
                         AND t1.entity_type_id = 3
                         AND t2.attribute_id = t1.attribute_id
                         AND t2.store_id = 0
                         AND t2.row_id = t0.row_id) AS url_path,
                     (SELECT `value`
-                       FROM eav_attribute t1, catalog_category_entity_int t2
+                       FROM ${table:eav_attribute} t1, ${table:catalog_category_entity_int} t2
                       WHERE t1.attribute_code = \'is_anchor\'
                         AND t1.entity_type_id = 3
                         AND t2.attribute_id = t1.attribute_id
                         AND t2.store_id = 0
                         AND t2.row_id = t0.row_id) AS is_anchor
-               FROM catalog_category_entity AS t0',
+               FROM ${table:catalog_category_entity} AS t0',
         SqlStatementKeys::CATEGORIES_BY_STORE_VIEW =>
             'SELECT t0.*,
                  IF (name_store.value_id > 0, name_store.value, name_default.value) AS name,
                  IF (url_key_store.value_id > 0, url_key_store.value, url_key_default.value) AS url_key,
                  IF (url_path_store.value_id > 0, url_path_store.value, url_path_default.value) AS url_path,
                  IF (is_anchor_store.value_id > 0, is_anchor_store.value, is_anchor_default.value) AS is_anchor
-               FROM catalog_category_entity AS t0
-          LEFT JOIN catalog_category_entity_varchar AS name_store
+               FROM ${table:catalog_category_entity} AS t0
+          LEFT JOIN ${table:catalog_category_entity_varchar} AS name_store
                  ON name_store.attribute_id = (
-                        SELECT attribute_id FROM eav_attribute
+                        SELECT attribute_id FROM ${table:eav_attribute}
                         WHERE attribute_code = \'name\' AND entity_type_id = 3
                     )
                     AND name_store.store_id = :store_id
                     AND name_store.row_id = t0.row_id
-          LEFT JOIN catalog_category_entity_varchar AS name_default
+          LEFT JOIN ${table:catalog_category_entity_varchar} AS name_default
                  ON name_default.attribute_id = (
-                        SELECT attribute_id FROM eav_attribute
+                        SELECT attribute_id FROM ${table:eav_attribute}
                         WHERE attribute_code = \'name\' AND entity_type_id = 3
                     )
                     AND name_default.store_id = 0
                     AND name_default.row_id = t0.row_id
-          LEFT JOIN catalog_category_entity_varchar AS url_key_store
+          LEFT JOIN ${table:catalog_category_entity_varchar} AS url_key_store
                  ON url_key_store.attribute_id = (
-                        SELECT attribute_id FROM eav_attribute
+                        SELECT attribute_id FROM ${table:eav_attribute}
                         WHERE attribute_code = \'url_key\' AND entity_type_id = 3
                     )
                     AND url_key_store.store_id = :store_id
                     AND url_key_store.row_id = t0.row_id
-          LEFT JOIN catalog_category_entity_varchar AS url_key_default
+          LEFT JOIN ${table:catalog_category_entity_varchar} AS url_key_default
                  ON url_key_default.attribute_id = (
-                        SELECT attribute_id FROM eav_attribute
+                        SELECT attribute_id FROM ${table:eav_attribute}
                         WHERE attribute_code = \'url_key\' AND entity_type_id = 3
                     )
                     AND url_key_default.store_id = 0
                     AND url_key_default.row_id = t0.row_id
-          LEFT JOIN catalog_category_entity_varchar AS url_path_store
+          LEFT JOIN ${table:catalog_category_entity_varchar} AS url_path_store
                  ON url_path_store.attribute_id = (
-                        SELECT attribute_id FROM eav_attribute
+                        SELECT attribute_id FROM ${table:eav_attribute}
                         WHERE attribute_code = \'url_path\' AND entity_type_id = 3
                     )
                     AND url_path_store.store_id = :store_id
                     AND url_path_store.row_id = t0.row_id
-          LEFT JOIN catalog_category_entity_varchar AS url_path_default
+          LEFT JOIN ${table:catalog_category_entity_varchar} AS url_path_default
                  ON url_path_default.attribute_id = (
-                        SELECT attribute_id FROM eav_attribute
+                        SELECT attribute_id FROM ${table:eav_attribute}
                         WHERE attribute_code = \'url_path\' AND entity_type_id = 3
                     )
                     AND url_path_default.store_id = 0
                     AND url_path_default.row_id = t0.row_id
-          LEFT JOIN catalog_category_entity_int AS is_anchor_store
+          LEFT JOIN ${table:catalog_category_entity_int} AS is_anchor_store
                  ON is_anchor_store.attribute_id = (
-                        SELECT attribute_id FROM eav_attribute
+                        SELECT attribute_id FROM ${table:eav_attribute}
                         WHERE attribute_code = \'is_anchor\' AND entity_type_id = 3
                     )
                     AND is_anchor_store.store_id = :store_id
                     AND is_anchor_store.row_id = t0.row_id
-          LEFT JOIN catalog_category_entity_int AS is_anchor_default
+          LEFT JOIN ${table:catalog_category_entity_int} AS is_anchor_default
                  ON is_anchor_default.attribute_id = (
-                        SELECT attribute_id FROM eav_attribute
+                        SELECT attribute_id FROM ${table:eav_attribute}
                         WHERE attribute_code = \'is_anchor\' AND entity_type_id = 3
                     )
                     AND is_anchor_default.store_id = 0
                     AND is_anchor_default.row_id = t0.row_id',
         SqlStatementKeys::CATEGORY_VARCHARS_BY_ENTITY_IDS =>
             'SELECT t1.*
-               FROM catalog_category_entity AS t0
-         INNER JOIN catalog_category_entity_varchar AS t1
+               FROM ${table:catalog_category_entity} AS t0
+         INNER JOIN ${table:catalog_category_entity_varchar} AS t1
                  ON t1.row_id = t0.row_id
-         INNER JOIN eav_attribute AS t2
+         INNER JOIN ${table:eav_attribute} AS t2
                  ON t2.entity_type_id = 3
                 AND t2.attribute_code = \'name\'
                 AND t1.attribute_id = t2.attribute_id
@@ -148,17 +148,17 @@ class SqlStatementRepository extends \TechDivision\Import\Repositories\SqlStatem
     );
 
     /**
-     * Initialize the the SQL statements.
+     * Initializes the SQL statement repository with the primary key and table prefix utility.
+     *
+     * @param \IteratorAggregate<\TechDivision\Import\Utils\SqlCompilerInterface> $compilers The array with the compiler instances
      */
-    public function __construct()
+    public function __construct(\IteratorAggregate $compilers)
     {
 
-        // call the parent constructor
-        parent::__construct();
+        // pass primary key + table prefix utility to parent instance
+        parent::__construct($compilers);
 
-        // merge the class statements
-        foreach ($this->statements as $key => $statement) {
-            $this->preparedStatements[$key] = $statement;
-        }
+        // compile the SQL statements
+        $this->compile($this->statements);
     }
 }